Extracting AyahHighlight and friends out of the HighlightingImageView
